Use a general crc32 implementation.
This simplifies the copyright management, as the old crc.{c,h}
had to be mentioned explicitly. This does NOT REALLY pull a new
dependency, as tgl already depends on -lz.
In fact, this ditches -lm, as the only usage (even including tgl)
was a single call to 'sqrt' from within crc32.c.